Refactoring to align with general practices
see https://github.com/golang-standards/project-layout#go-directories
golang layout standards
