+++ pkg/bowline/bowline_test.go
@@ -1,4 +1,4 @@
-package main
+package bowline
 
 import (
 	"os"
