+++ pkg/compose/compose_test.go
